\name{iteRates-package}
\alias{iteRates-package}
\alias{iteRates}
\docType{package}
\title{
iteRates
}
\description{
Iterates through a phylogenetic tree to identify regions of rate variation using the parametric rate comparison test.
}
\details{
\tabular{ll}{
Package: \tab iteRates\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2011-05-24\cr
License: \tab GPL 3.0\cr
LazyLoad: \tab yes\cr
}
The user provides a phylogenetic tree of object class phylo. The package will iterate through all useable subtrees and identify regions of the tree with different rates of diversification using the parametric rate comparison test.
}
\author{
Premal Shah, Benjamin Fitzpatrick and James Fordyce.

Maintainer:
James A. Fordyce <jfordyce@utk.edu>
}

